Arff Reader
Arff Reader was implemented ... at least partially. It would be great if
the format of the arff is followed exactly and arff has its all parts,
including some @data. It is tested and working as expected for its
state.

It is added also attribute index, so the search for attribute's index
should be faster.